+++ src/js/export.js
@@ -2,5 +2,12 @@ import HomePage from './components/home-page'
 import ListingItemsPage from './components/list-items-page'
 import ListingItemPage from './components/list-item-page'
 import DocsPage from './components/docs-page'
+import CheckoutPage from './components/checkout-page'
 
-export { HomePage, ListingItemsPage, ListingItemPage, DocsPage }
+export {
+  HomePage,
+  ListingItemsPage,
+  ListingItemPage,
+  DocsPage,
+  CheckoutPage
+}